Move CashRegister class to new namespace
